source_manager/requirements.txt
Flask~=2.3.2
tweepy~=4.14.0
python-dotenv~=1.0.0
Unidecode~=1.3.6